+++ src/Domain/User/Repository/UserEventStoreInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Domain\User\Repository;
 
-use App\Domain\User\AggregateRoot\User;
+use App\Domain\User\Entity\User;
 use Ramsey\Uuid\UuidInterface;
 
 interface UserEventStoreInterface